don't expose secstore secrets in /mnt/term/dev/secstore
this feature is too dangerous. i can imagine it being there
so a server side factotum can use the drawterms secstore
keys that it already loaded, but this makes it just too easy
to steal the keys.

now we zero the secstorebuf after we extracted the plan9
key from it so no secrets are left behind in drawterms
memory after authentication is done.