fix devfs-*.c to not use cname hack to get file path, get rid of MAXPATH limit, make root accessible under /mnt/term/root
under linux, it was not possible to access the clients /mnt directory as it
gets hidden by #/mnt, to solve this, we first bind #U to /root, and then
let cpumain() bind any subdirectory (-r option) in there and finally bind
/root after /. for this to work, i had to get rid of the dreadfull cname hack,
where the c->name is used to derive the unix file path. this did not work when
the device was accessed form a different bind... and it also leaked memory in
cclone() as the c->name's where doubly maintained in fswalk().

also get rid of the fixed MAXPATH sized stack buffers. each devfs now maintains
a dynamic buffer in ther Ufsinfo structure. under windows, this is a wide string.

remove the #ifdef UNICODE maze from devfs-posix.c, we always assume unicode
now with wide strings.