screen: add screensize() function to reallocate the gscreen
previously, we where allocating the gscreen once with the
maximum screen size as its dimensions and then adjust
gscreen->clipr for the window size.

this is simple, but doesnt play well with multi-monitor
systems where there is no maximum screen size.
so instead, we reallocate the gscreen image to the actual
window size using the new screensize() function.

screensize() can be a no-op (x11), then gscreen->clipr
is adjusted just as before.

when screensize() reallocated the screen, then it should
set gscreen->clipr to the zero rect.