gui-metal: request screen update again in 8 ms
This is still a work around for the bug that texture update on
the CPU side are not properly synchronized with the display
update on the GPU side.

This change should minimize the display latency while still
provides a second update that is hopefully contains the
complete CPU memory update.