remove cpu specific _sha1block() and _md5block() functions
we used to have cpu specific _sha1block() and _md5block()
functions for 386, but every other arch just copied the
portable c version.

md5() and sha1() are not really on the critical path as
we use chacha20/poly1305 over TLS for communication with
the cpu server.

so just get rid of the noise.