gui-wl: dont bind more then one seat
We dont currently have any logic for handling multiseat, so we should
avoid binding more then one seat. This locks us to the first seat, which is
often the default. This also moves the data managers binding over to the
seat binding, these managers are scoped per seat so our binding logic
should reflect that.