devfs: create handle OEXCL, mask permissions with parent directory, fix leaks
OEXCL has not been handled correctly. We have to
return an error if the file already exists.

The file-permissions should only be applied when
we create a new file.

When creating a new file, we mask te permissions
with the permissions of the parent directory
(see open(5)).

There where a few potential leaks. After creating
file file, we do a stat to get the qid information,
but this can potentially fail, eigther leaving the
Chan in the wrong state or leaking memory. Fix it.