import drawlock fixes from 9front (thanks foura)
These were first found in drawterm commit
54f615f6c0090c7c0011f4121ecf54fcf1e07a6b. In which due to the
locking issues a race was found that crashed the wayland backend.
This turned out to be a bug in devdraw itself and was fixed in 9front
commit e7d03d19100fe3180906c8ce3f2cab3ba9387fe8, in the meantime the
temporary hack in drawterm was reverted. The proper solution was merged
as mentioned but not propigated back to drawterm until now.

This came up again as it was found in some instalations of KDE's
wayland to cause crashes more reliably in resizes.