gui-wl: fix more issues surfaced by cage
* We were calling wlresize() in the csd handler, this
was initially done to ensure that we draw the csd on
the first blit. This can instead be satisfied by doing
a roundtrip during initialization and being more careful
about the first blit in wlattach().

* Free the gscreen image on resize

* On cage since the "monitor" is the window,
we had a bug where x and y must both change
for us to consider the monitor to be larger.

* screenwin() should handle gscreen->clip where
r.min is not (0, 0).

* Refactor termscreenputs to avoid this canqlock
bussiness. When resizing this would cause an issue
where we would fail redraw the kmesg buffer.