gui-wl: Correct the previous cursor fix
The Plan 9 cursor bitmap is transformed into an ARGB8888 surface to be displayed by the
Wayland compositor. However, the mask used to decide the pixel colour was incorrect.

When j = 0 the mask is 1<<16 which is not valid.
When j = 15 the mask is 1<<1 which is not the final pixel.
This results in the cursor image being shifted 1px right, off the edge of the surface.
The correct mask is (1<<15) >> j;

Thanks moody for cleaning up the patch.